Fix float type checks, SSH agent & tooling updates
Correct various type checks to use Float.class for float handling across registry, metadata, API scanner, servlet and samples. Revamp ShellUseAgent SSH setup: support key auth, set SSH configs and keep-alive, stream command output in real-time, store session/channel in global data, and pass session to ssh_tool; extend AuthInfo with keyPath. Update DemoPlanAndSolve to choose MCP client commands based on OS, remove get_current_date from demo executor tools, and adjust demo ShellUseAgent auth construction. Add two .oxygent SKILL.md files (framework explorer and weather), remove /.oxygent from .gitignore, add improved error logging in StdioMCPClient, and remove the multimodal toContent method from ObservationData. Misc formatting and minor logging/refactor tweaks.

+++ oxygent-core/src/main/java/com/jd/oxygent/core/oxygent/mcpservers/engine/MCPToolRegistry.java
@@ -173,7 +173,7 @@ public class MCPToolRegistry {
                 return Double.parseDouble(defaultValue);
             if (targetType == Boolean.class || targetType == boolean.class)
                 return Boolean.parseBoolean(defaultValue);
-            if (targetType == Double.class || targetType == float.class)
+            if (targetType == Float.class || targetType == float.class)
                 return Double.parseDouble(defaultValue);
             if (targetType == Long.class || targetType == long.class)
                 return Long.parseLong(defaultValue);
@@ -197,7 +197,7 @@ public class MCPToolRegistry {
                 return Integer.parseInt(stringValue);
             if (targetType == Double.class || targetType == double.class)
                 return Double.parseDouble(stringValue);
-            if (targetType == Double.class || targetType == float.class)
+            if (targetType == Float.class || targetType == float.class)
                 return Double.parseDouble(stringValue);
             if (targetType == Long.class || targetType == long.class)
                 return Long.parseLong(stringValue);
